back/requirements.txt
Flask==2.1.1
Flask-Cors==3.0.10
google-api-core==2.7.2
google-auth==2.6.4
google-cloud-bigquery==3.0.1
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0